refactor rule source tests
